commit para pre merge

+++ app/src/main/java/com/example/espacios_um/modelos/Usuario.java
@@ -25,7 +25,7 @@ public class Usuario implements Serializable {
 
     }
 
-    public int getID() {
+    public long getID() {
         return id;
     }
